implementacion de test unitario y test instrumental

+++ app/src/main/java/com/example/superheroes/data/local/SuperheroEntity.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 
 @Entity(tableName = "tabla_superheroes")
 data class SuperheroEntity(
-    @PrimaryKey(autoGenerate = true) var id: Long = 0,
+    @PrimaryKey(autoGenerate = true) var id: Int = 0,
     val nombre: String,
     val origen: String,
     @SerializedName("imagenLink") val imagen: String,
